requirements.txt
anvil-uplink==0.3.41
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
Django==4.0
dnspython==2.1.0
future==0.18.2
idna==3.3
image==1.5.33
Jinja2==3.0.3
MarkupSafe==2.0.1
mime==0.1.0
mongoengine==0.23.1
mpmath==1.2.1
numpy==1.21.4
pandas==1.3.4
passlib==1.7.4
Pillow==8.4.0
psutil==5.8.0
pwinput==1.0.2
pycparser==2.21
pymongo==3.12.2
pymongocrypt==1.2.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
sympy==1.9
tzdata==2021.5
urllib3==1.26.7
ws4py==0.5.1
